requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
pymysql==1.1.0

# Vector Database
chromadb==0.4.17

# LLM Integration
openai>=1.30.0

# Environment Variables
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.24.3

# HTTP Client for API calls
httpx==0.25.2

# MySQL 8 auth plugin support
cryptography>=41.0.0

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0